Implement create_fermi_isoptropic_diffuse_model helper function
